pyproject.toml
[tool.poetry]
name = "climatoology"
version = "2.0"
description = "This package provides the underlying functionality to implement Climate Action Operators and Plugins."
authors = [
  "Maciej Adamiak <maciej.adamiak@heigit.org>",
  "Moritz Schott <moritz.schott@heigit.org>"
]
readme = "README.md"
repository = "https://gitlab.gistools.geog.uni-heidelberg.de/climate-action/climatoology"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "~3.11.5"
pika = "^1.3.2"
pydantic = "^2.3.0"
bibtexparser = "^1.4.0"
semver = "^3.0.1"
minio = "^7.1.16"
rasterio = "^1.3.8"
tqdm = "^4.66.1"
requests = "^2.31.0"
pillow = "^10.0.0"
amqpstorm = "^2.10.6"
cachetools = "^5.3.1"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
websocket = "^0.2.1"
hydra-core = "^1.3.2"
email-validator = "^2.0.0.post2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
